refactor : reviews and add animations
