+++ src/components/homePage/testimonialStatic.tsx
@@ -1,55 +1,6 @@
 import Image from "next/image";
-import google from "../../../public/award-logo-1.svg";
-import trustpilot from "../../../public/trustpilot-logo.png";
-import user1 from "../../../public/user1.png";
-import user2 from "../../../public/user2.png";
-import user3 from "../../../public/user3.png";
 import Link from "next/link";
-
-const testimonials = [
-  {
-    name: "John Doe",
-    title: "CTO, Nugen IT Services",
-    quote:
-      "Nugen's team has been a valuable partner in developing scalable, secure IT solutions. Their technical expertise is unmatched.",
-    image: user1,
-  },
-  {
-    name: "Sarah Lee",
-    title: "Product Manager, TechVentures",
-    quote:
-      "Nugen's agile approach helped us launch our product on time with impressive results.",
-    image: user3,
-  },
-  {
-    name: "David Smith",
-    title: "Lead Developer, SoftwareCorp",
-    quote:
-      "The custom solutions provided by Nugen helped us improve system efficiency and performance significantly.",
-    image: user2,
-  },
-  {
-    name: "Linda Tan",
-    title: "CEO, InTech Solutions",
-    quote:
-      "We saw a remarkable improvement in our system's security after working with Nugen.",
-    image: user3,
-  },
-  {
-    name: "James Carter",
-    title: "Senior Developer, DataSolutions",
-    quote:
-      "Nugen's tech support has been essential in overcoming the most complex challenges.",
-    image: user1,
-  },
-  {
-    name: "Mary Wilson",
-    title: "Director, CloudTech",
-    quote:
-      "Their approach to cloud migration and management has brought us significant cost savings.",
-    image: user2,
-  },
-];
+import { TESTIMONOAL_DATA_ONE } from "@/src/json/reviews/reviews";
 
 interface TestimonialStaticProps {
   className?: string;
@@ -59,7 +10,9 @@ export default function TestimonialStatic({
   className,
 }: TestimonialStaticProps) {
   return (
-    <section className={`py-16  text-black bg-gradient-to-br from-white via-[#fff9f7] to-[#ffece4] ${className}`}>
+    <section
+      className={`py-16 text-black bg-gradient-to-br from-white via-[#fff9f7] to-[#ffece4] ${className}`}
+    >
       <div className="text-center mb-10">
         <div className="text-yellow-400 text-3xl mb-2">★★★★★</div>
         <h2 className="text-3xl font-semibold">What Our Students Say</h2>
@@ -71,22 +24,22 @@ export default function TestimonialStatic({
 
       {/* Testimonials Cards */}
       <div className="grid md:grid-cols-3 gap-8 px-6 max-w-7xl mx-auto">
-        {testimonials.map((t, i) => (
+        {TESTIMONOAL_DATA_ONE.map((t, i) => (
           <div
             key={i}
             className="bg-white p-6 rounded-xl shadow-lg hover:shadow-2xl transform transition-all duration-300 relative"
           >
             <div className="flex items-center gap-3 mb-4">
               <div className="w-12 h-12 rounded-full overflow-hidden border-2 border-[#ff6b3d]">
-                <Image src={t.image} alt={t.name} width={48} height={48} />
+                <Image src={t.imageURL} alt={t.name} width={48} height={48} />
               </div>
               <div>
                 <p className="font-semibold text-lg">{t.name}</p>
-                <p className="text-xs text-black">{t.title}</p>
+                <p className="text-xs text-black capitalize">{t.source}</p>
               </div>
             </div>
             <p className="text-md italic text-[#2b2b2b]">
-              &quot;{t.quote}&quot;
+              &quot;{t.text}&quot;
             </p>
             <div className="text-[#ff6b3d] text-3xl absolute bottom-4 right-4">
               ”
@@ -104,8 +57,13 @@ export default function TestimonialStatic({
           Read More Reviews
         </Link>
         <div className="flex justify-center items-center gap-6 mt-10 flex-wrap">
-          <Image src={google} alt="Google" width={100} height={35} />
-          <Image src={trustpilot} alt="Trustpilot" width={100} height={35} />
+          <Image src="/award-logo-1.svg" alt="Google" width={100} height={35} />
+          <Image
+            src="/trustpilot-logo.png"
+            alt="Trustpilot"
+            width={100}
+            height={35}
+          />
         </div>
       </div>
     </section>
